requirements.txt
fastapi>=0.110.0
uvicorn==0.24.0
pydantic>=2.6.1
pyautogen==0.9.1a1
python-dotenv==1.0.0
python-multipart==0.0.6
asyncio-mqtt==0.16.1
aiofiles==23.2.1
loguru==0.7.2
typing-extensions==4.8.0
aiohttp==3.9.1 